+++ src/types/qwlayershellv1.h
@@ -31,8 +31,6 @@ public:
     QW_FUNC_MEMBER(layer_surface_v1, surface_at, wlr_surface *, double sx, double sy, double *sub_x, double *sub_y)
     QW_FUNC_MEMBER(layer_surface_v1, popup_surface_at, wlr_surface *, double sx, double sy, double *sub_x, double *sub_y)
     QW_FUNC_MEMBER(layer_surface_v1, configure, uint32_t, uint32_t width, uint32_t height);
-
-protected:
     QW_FUNC_MEMBER(layer_surface_v1, destroy, void)
 };
